Text Heuristics for identifying data type
I made some progress in identifying the type of the tokens in some text (be it query or otherwise) by fleshing out some of the functions. The goal is to divide each token into a type (I, II, III, or IV), which can then be saved to the extractor and retrieved by the caller.

In so doing, I needed to make a list of units of measure. The list is unfinished, but I started the file.