SVM and Data
I made two big changes.
The most important one is about the data that we are using. I was able to make some small changes to finalize file loading in the load_sql.py file. Now, all of the data categories should be loading.
There were some changes needed in the data files themselves. There were some unneeded columns in some files. In the DataScientist file, strings were escaped by "" instead of \". For simplicity, they were just changed to be apostrophes. The data files should be relatively final now. We may want to change the order of entries for the sake of diversity, and make some unit conversions, but other than that, they should be done.

In regards to the SVM, I think that the idea is all set up as demonstrated in svm_test.py. Now I just need the tagged data of queries to actually put it to use.