Performance and Aesthetic Database Changes
I made some very small changes to the dataset files and how the database is saved. I added spaces to make the SQL commands easier to read, and I removed the rowId for tables that already had a primary key set up. These changes should not affect any functionality.
I also noticed that the dataset file for jobs included the company ranking (again) in the company name. I started removing the most common instances of this for clarity and because it may help in processing downstream.