Prep for SQL Building
I made some necessary stub files for Type II so that they can be filled in when we get the data. Until then, the process can operate (just not properly typify type II for the missing categories).

There was a small change made to the order in typifying. Some car and motorcycle models are numbers. Therefore, the program will try a token against all tries first before checking if it is a number (in which case it is sorted to Type III).