Type I where clause
I worked on the main program to generate a Type I where clause. The program will extract Type I values, find which columns they apply to in the table, merge where possible, then return a list of substring matches that the sought product should fulfill. These are intentionally kept separate in the list so any number can be dropped if necessary to fulfill a partial match.

The logic for Type II should be pretty much the same. I will need to abstract some features and manually enter the Type II columns, but the searching and combining is common to both types.

The Type III logic will be very different since it will be focused on ranges instead of simple substring matches.