Robust Dataset Loading
I found an entry that was ill-formatted in the data scientist jobs file, which was then corrected. In the process, I made the entry loading procedure slightly more robust.

Also, in an earlier update, the file for motorcycles was changed to include a hyphen instead of a space. This was fixed in the loading script.