Type 3 Exact matching work
Type III exact matching is not yet complete, but I got a good start on it. I figured that I could put the unit information in the table for each relevant attribute. Then in the Type III building, I can search the relevant table looking for unit types to match to what I find for the unit of the numbers pulled out of the query.

It is possible for the units of several attributes in the same table to be the same. This is potentially problematic, so I will want to put an OR between each of the possible attributes for each single value (it is not useful to consider it as an AND, otherwise the user would indicate by listing it two times. [We don't currently check for such a redundancy, but perhaps we should!])