Type I/II and more work on Type III
I did some work on Type I/II exact matching. I was able to change the dataset loading to include a space before and after each entry. Therefore, when I am doing substring matching, I can *always* match with a space before and after to know that the token appears as a distinct word in the search.

I was also able to make some nondeterminism of sorts in query interpretation of Type I and Type II values. If a token is reused in two different searches, it will generally (the system is not 100% accurate) place the two searches in an OR statement with each other. Previously, it would put them adjacent in the list, to be interpreted as AND clauses.

Also, I did a little bit of work on Type III exact matching. Currently the program only performs equivalency for Type III values found (ignoring any "less than", "greater than", etc prefixes). However, that will be changed in a later update.