Boundary Operations
I finished work on simple boundary operations such as >, >=, <, and <=. In a real query, these probably occur in words such as "at least", "or more", "no more than", or "less than", for example.  Therefore, these word boundaries are converted to their symbol counterpart for simplicity then handled.

I still need to make ranges in the query work for type 3 and ranges in the data (which is not currently in place).