Range Simplification
I wrote some simple code in what was the boundQuery function (and is now the standardizeQuery function) to simplify common range phrases, such as "between * and *", "from * to *", and "*-*".

I also had to make a small edit to the tokenization process to properly catch ranges, which sometimes are expressed without spaces (fx $20-30).