ConstraintBuilder's init
I was able to offload a lot of the work performed in the query analysis to the init function. This way, the same object can be reused for multiple different queries if desired (without needing to reload all the files and constants between query analysis).

I also made a small change to the table object structure (no changes to the database structure itself) in anticipation of an update to superlatives and order by.

It may also be worth mentioning that I removed the redundancy that existed from specifying the border words again in the main.py file. Now both use cases use the bounders saved from boundary-synonyms.txt.