Superlative ordering
Following a similar path of what I did with bounding (less than, greater than), I implemented superlative handling. The program standardizes complete and incomplete superlative phrases into either << or >> (with some unit or attribute being affected). Then, in the orderBy function, the program looks for the standardized tokens, matches the unit or attribute to a corresponding attribute in the table context, and returns a list of all found.