Small changes
I made some small changes to the program. I made a separate function for spelling correction, I changed the range code to use BETWEEN syntax (since it is easier for the database to optimize), and I changed the program to take the query as a command line argument instead of using a hard-coded value. This last change should result in fewer unimportant changes since testing different queries will be outside the repository files.