Partial Ordering and Bug Fixing
I implemented partial ordering, such that when a user provides some boundary (such as < x cost), then it is assumed partial matches should be preferred with a low price. It is assumed that the first listed bounds are more important than later ones.

I also fixed a couple bugs in regards to loading the database. These went unnoticed since I had not tried to reload the database for a few updates.