Explicit Logical Operators
I made some big changes. First and foremost, I made some organizational changes. I removed the extract.py file that we have not been using for a while now. I tried to move some of the logic out of main.py to clean up the file, and as a result, I made the standard.py file to hold the Standardizer class, and I put the new OperatorHandler class in opeval.py (since it fit nicely there). This accounts for nearly all the deletions and a significant amount of the additions.

However, I also made some improvements to the program itself. In a previous update, I started on explicit logical operators in the query (not, and, or). Now I was able to finish up what I was doing and present a first draft that runs end-to-end, though may not be bug free.

Most of the changes related to this were in the extractOperated method in main.py. The method works through the return of operation evaluation, and goes through the terms (recursively in the case of OperatorRelations) to flatten all constraints. This flattened constraint list (of the three types) is returned back to fromQuery. Due to the recursive nature of the process, fromQuery can print out the returned constraints, find the ordering, and send the data to the PartialMatcher.