Column title instead of unit
Sometimes users indicate the column of the Type III data through the column rather than the unit. For example, someone could say "500 miles" which would be assumed as the mileage, or they could say "mileage 500", which has the same meaning.
This thinking is now usable for the program. Type II column names are handled as "units" where no other unit is provided for a Type III numeric value.