Fixed bug for joint units run-on
There was a small bug (which now has been resolved), in which several unit words in a row would be concatenated, and even though not all were used, all would be blacklisted. This was most common in cases such as "500 miles $ 1000" with a post-unit followed by a pre-unit and without any punctuation or separator.