Better typification grouping
With some minor modifications to the trie, I was able to make typification and type grouping much more clever. If the word is part of some word in the type's trie, the next word is concatenated on with a space, and that is tried. If it works, then the words are grouped together for that type. If the word exceeds the length of the words in the trie for the pattern, the process is automatically aborted.