Complex Explicit Boolean Operators
I changed how explicit boolean operators are processed, where the types on both sides of the operator are attempted to match, but may fail if not within a given tolerance. This means that "honda or red toyota" would be processed as "honda or (red toyota)", since toyota and honda match types.