Content-based partial matching
I am working on implementing content-based partial matching. The ground work is all present, though I need to work on a couple of the graph files (namely, material and work-sector), and I could perhaps make more graph files to use.

I was about to seriously modify the partial matching scheduler, but I figured that it would be wise to save first with the program in a working state. I anticipate that next commit I will have connected my work in content_match.py with partial.py.

As a side note, I made some small changes to the database file such that it would use Domains. I also modified the Operation Relations to account for Type II or Type III modifiers on a Type I noun (as adjectives and modifiers come before the noun in English).