Edge Weights and Type III balancing
I made some tweaks to the content matching. First of all, I increased the penalty of incorrect years (since for cars and motorcycles, the exact year may determine how the car looks). I also modified the graph subsystem such that each edge can have a weight (the default is 1). Then in the ranking, there is a 1:5 ratio of edge weight to penalty to score. This ratio may need to be tweaked some more.