Exponential decay similarity scoring model
I implemented an exponential decay similarity scoring model. I still need to test and perhaps modify the constants, but I knew that I needed a different model than the linear model we were using before.
In the previous linear model, the distance of similarity was linear. For example, if the target is 0, then the distance between 0 and 1 was equivalent to the distance between 1 and 2. This is problematic for two reasons:
1. The linear model necessarily has a cap. There is some point, no matter the scale, that increasing distance will not result in less similarity. This is an issue if the scale is very small, which can often occur with small target values (since the target value is an easy metric for scale).
2. It goes against reason that if a user is looking for a car made in 2010, that the difference between a car made in 2009 and 2008 would be equivalent to the difference between a car made in 2000 and 1999. The further the distance is from the target value, the less each additional year (or whatever unit) matters. This would not be a problem in a void if only year was being measured, but in a system that compares many requirements, the linear model yields undesirable results.