Bound Use Cases, Spelling Disable flag
I gave bounds use cases (they can be used before the value, after the value, or either before or after. The majority of bounds may only be used before the unit.

I modified the precedence of the different explicit boolean operators. This was a small tweak, but it exposed that the system is fairly brittle. The opeval set-up needs an overhaul.

I added a new flag for disabling spelling correction. This speeds up regular running several times.