Bug fixing and type tweaking
I fixed some bugs in regards to partial ranking, so the system should be more thorough now (though the tradeoff is it is likely slower).
I changed some of the type categorizations of some queries that I worked with. Type II still needs to be cleaned up extensively, so I imagine there will be many similar changes in the future.